requirements.txt
# RAG Chatbot Dependencies
streamlit>=1.28.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.6.0
pypdf>=3.17.0
python-dotenv>=1.0.0
